% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dwca_tables.R
\name{build_dwca_tables}
\alias{build_dwca_tables}
\title{Build Darwin Core Archive tables from input data and field mapping}
\usage{
build_dwca_tables(
  df,
  dwc_terms,
  target_database = "GBIF",
  selected_terms = NULL,
  emof_spec = NULL
)
}
\description{
Build Darwin Core Archive tables from input data and field mapping
}
