% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dwc_pipeline.R
\name{clean_dwc_pipeline}
\alias{clean_dwc_pipeline}
\title{Clean and validate a mapped Darwin Core data.frame}
\usage{
clean_dwc_pipeline(
  df,
  coord_epsg_in = 4326,
  target_epsg = 4326,
  force_parzer = TRUE
)
}
\arguments{
\item{df}{data.frame after mapping (DwC column names)}
}
\value{
list(data=cleaned_df, issues=issues_df, summary=summary_df)
}
\description{
Clean and validate a mapped Darwin Core data.frame
}
