% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dwc_mapping.R
\name{mod_dwc_mapping_server}
\alias{mod_dwc_mapping_server}
\title{Darwin Core mapping module server}
\usage{
mod_dwc_mapping_server(id, df_in)
}
\arguments{
\item{id}{Module id.}

\item{df_in}{Reactive returning a data.frame (use ingest$tidy).}
}
\value{
A list of reactives: mapping, mapped, cleaned, issues, clean_summary, validation.
}
\description{
Darwin Core mapping module server
}
