% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dwc_mapping.R
\name{mod_dwc_mapping_ui}
\alias{mod_dwc_mapping_ui}
\title{Darwin Core mapping module UI}
\usage{
mod_dwc_mapping_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
A Shiny UI definition.
}
\description{
Shiny module UI for mapping user columns to Darwin Core terms.
}
