% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_emof_data_editor.R
\name{mod_emof_data_editor_server}
\alias{mod_emof_data_editor_server}
\title{eMoF Data Editor server}
\usage{
mod_emof_data_editor_server(id, emof_in)
}
\arguments{
\item{id}{Module id.}

\item{emof_in}{Reactive returning the eMoF table produced upstream.}
}
\value{
A list with reactives: emof, lookup, ready.
}
\description{
eMoF Data Editor server
}
