% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_emof_data_editor.R
\name{mod_emof_data_editor_ui}
\alias{mod_emof_data_editor_ui}
\title{eMoF Data Editor UI}
\usage{
mod_emof_data_editor_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
eMoF Data Editor UI
}
