% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_homepage.R
\name{mod_homepage_server}
\alias{mod_homepage_server}
\title{Homepage module server}
\usage{
mod_homepage_server(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Module id.}

\item{session}{Shiny session object.}
}
\value{
Module server function.
}
\description{
Homepage module server
}
