% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_homepage.R
\name{mod_homepage_ui}
\alias{mod_homepage_ui}
\title{Homepage module UI}
\usage{
mod_homepage_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
A Shiny UI definition.
}
\description{
Homepage module UI
}
