% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_identification_cleaning.R
\name{mod_identification_cleaning_server}
\alias{mod_identification_cleaning_server}
\title{Identification Data Cleaning server}
\usage{
mod_identification_cleaning_server(id, df_in, mapping_in)
}
\arguments{
\item{id}{Module id.}

\item{df_in}{Reactive returning the cleaned output from Field Mapping.}

\item{mapping_in}{Reactive returning the Field Mapping table.}
}
\value{
A list with reactives: df_out, lookup, ready, source_column.
}
\description{
Identification Data Cleaning server
}
