% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_identification_cleaning.R
\name{mod_identification_cleaning_ui}
\alias{mod_identification_cleaning_ui}
\title{Identification Data Cleaning UI}
\usage{
mod_identification_cleaning_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Identification Data Cleaning UI
}
