% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ingest.R
\name{mod_ingest_server}
\alias{mod_ingest_server}
\title{Ingestion module server}
\usage{
mod_ingest_server(id, example_map)
}
\arguments{
\item{id}{Module id.}

\item{example_map}{Named list mapping example labels to file paths.}
}
\value{
A list of reactives: raw, tidy, current, ready, label, dropped_log.
}
\description{
Ingestion module server
}
