% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ingest.R
\name{mod_ingest_ui}
\alias{mod_ingest_ui}
\title{Ingestion module UI}
\usage{
mod_ingest_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\value{
A Shiny UI definition.
}
\description{
Shiny module UI for ingesting data from upload or examples.
}
