% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metadata.R
\name{mod_metadata_server}
\alias{mod_metadata_server}
\title{Metadata module server}
\usage{
mod_metadata_server(id)
}
\arguments{
\item{id}{Module id}
}
\value{
A list of reactives with metadata content
}
\description{
Metadata module server
}
