% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qc.R
\name{mod_qc_server}
\alias{mod_qc_server}
\title{QC page server}
\usage{
mod_qc_server(id, event_in, occ_in, emof_in = NULL, pre_issues_in = NULL)
}
\arguments{
\item{id}{Module id}

\item{event_in}{Reactive returning event table}

\item{occ_in}{Reactive returning occurrence table}

\item{emof_in}{Reactive returning emof table (may be NULL or empty)}

\item{pre_issues_in}{Reactive returning pre-split issues (optional)}
}
\value{
list with qc_res reactive and can_export reactive
}
\description{
QC page server
}
