% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxonomy_match.R
\name{mod_taxonomy_match_server}
\alias{mod_taxonomy_match_server}
\title{Taxonomy match server}
\usage{
mod_taxonomy_match_server(id, df_in)
}
\arguments{
\item{id}{module id}

\item{df_in}{reactive df after DwC mapping}
}
\value{
list(df_out, lookup, issues, ready)
}
\description{
Taxonomy match server
}
