% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxonomy_match.R, R/teste.R, R/testes_2.R
\name{mod_taxonomy_match_ui}
\alias{mod_taxonomy_match_ui}
\title{Taxonomy match UI}
\usage{
mod_taxonomy_match_ui(id)

mod_taxonomy_match_ui(id)

mod_taxonomy_match_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Taxonomy match UI

Taxonomy match UI

Taxonomy match UI
}
