% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dwc_pipeline.R
\name{clean_dwc_pipeline}
\alias{clean_dwc_pipeline}
\title{Clean and validate a mapped Darwin Core data.frame}
\usage{
clean_dwc_pipeline(
  df,
  coord_epsg_in = 4326,
  target_epsg = 4326,
  force_parzer = TRUE,
  bathy = NULL,
  bathy_auto_download = TRUE,
  bathy_depth_tolerance_m = 200,
  swap_auto_fix = FALSE,
  preserve_original_coords = FALSE,
  create_geodetic_datum = FALSE,
  geodetic_datum_value = "WGS84",
  standardize_event_date = TRUE
)
}
\arguments{
\item{df}{data.frame after mapping (DwC column names)}

\item{coord_epsg_in}{EPSG code of input coordinates}

\item{target_epsg}{EPSG code to reproject to (default 4326)}

\item{force_parzer}{logical; if TRUE, try parzer parsing even when values look numeric}

\item{bathy}{optional marmap bathy object already loaded/downloaded}

\item{bathy_auto_download}{logical; if TRUE and bathy is NULL, tries to fetch bathymetry with marmap for the dataset extent}

\item{bathy_depth_tolerance_m}{numeric; tolerance (meters) to compare reported depth with bathymetry}

\item{swap_auto_fix}{logical; if TRUE, auto-swap coordinates in clear range-based cases}

\item{preserve_original_coords}{logical; if TRUE, preserve original decimalLatitude/decimalLongitude
into verbatimLatitude/verbatimLongitude before any parsing/reprojection, when those
verbatim fields do not already exist}

\item{create_geodetic_datum}{logical; if TRUE, create/fill geodeticDatum}

\item{geodetic_datum_value}{character; value used to fill geodeticDatum when enabled}

\item{standardize_event_date}{logical; if TRUE, try to normalize eventDate to ISO-8601}
}
\value{
list(data=cleaned_df, issues=issues_df, summary=summary_df)
}
\description{
Clean and validate a mapped Darwin Core data.frame
}
