% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qc.R
\name{mod_qc_server}
\alias{mod_qc_server}
\title{QC page server}
\usage{
mod_qc_server(id, event_in, occ_in, emof_in = NULL, pre_issues_in = NULL)
}
\arguments{
\item{id}{Module ID.}

\item{event_in}{Reactive returning event table.}

\item{occ_in}{Reactive returning occurrence table.}

\item{emof_in}{Reactive returning eMoF table, may be NULL.}

\item{pre_issues_in}{Reactive returning pre-existing issues, may be NULL.}
}
\value{
A list with \code{qc_res} and \code{can_export} reactives.
}
\description{
QC page server
}
